chore: extract parameterization to client-engine-runtime
