+++ packages/client-common/package.json
@@ -24,7 +24,6 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@prisma/client-engine-runtime": "workspace:*",
     "@prisma/dmmf": "workspace:*",
     "@prisma/driver-adapter-utils": "workspace:*",
     "@prisma/generator": "workspace:*",
