+++ packages/client-common/src/QueryCompiler.ts
@@ -1,9 +1,8 @@
-import type { BatchResponse } from '@prisma/client-engine-runtime'
 import { ConnectionInfo, Provider } from '@prisma/driver-adapter-utils'
 
 export type QueryCompiler = {
   compile(request: string): {}
-  compileBatch(batchRequest: string): BatchResponse
+  compileBatch(batchRequest: string): {}
   free(): void
 }
