+++ packages/client-engine-runtime/src/parameterization/parameterize-tests/filter-operators.test.ts
@@ -1,13 +1,14 @@
 import type { JsonQuery } from '@prisma/json-protocol'
 import type { ParamGraph } from '@prisma/param-graph'
+import { beforeAll, describe, expect, it } from 'vitest'
 
 import { parameterizeQuery } from '../parameterize'
 import { getParamGraph } from './test-fixtures'
 
 let paramGraph: ParamGraph
 
-beforeAll(async () => {
-  paramGraph = await getParamGraph()
+beforeAll(() => {
+  paramGraph = getParamGraph()
 })
 
 describe('parameterizeQuery filter operators', () => {
