+++ packages/client-generator-js/src/getDMMF.ts
@@ -1,15 +1,1 @@
-import type * as DMMF from '@prisma/dmmf'
-import type { GetDMMFOptions } from '@prisma/internals'
-import { getDMMF as getRawDMMF } from '@prisma/internals'
-
-import { externalToInternalDmmf } from './externalToInternalDmmf'
-
-export function getPrismaClientDMMF(dmmf: DMMF.Document): DMMF.Document {
-  return externalToInternalDmmf(dmmf)
-}
-
-// Mostly used for tests
-export async function getDMMF(options: GetDMMFOptions): Promise<DMMF.Document> {
-  const dmmf = await getRawDMMF(options)
-  return getPrismaClientDMMF(dmmf)
-}
+export { getInternalDMMF as getDMMF, externalToInternalDmmf as getPrismaClientDMMF } from '@prisma/internals'
