+++ packages/client-generator-js/src/index.ts
@@ -1,4 +1,3 @@
-export { externalToInternalDmmf } from './externalToInternalDmmf'
 export { generateClient, type GenerateClientOptions } from './generateClient'
 export { PrismaClientJsGenerator } from './generator'
 export { getDMMF } from './getDMMF'
