+++ packages/client-generator-ts/src/index.ts
@@ -1,4 +1,3 @@
-export { externalToInternalDmmf } from './externalToInternalDmmf'
 export { generateClient, type GenerateClientOptions } from './generateClient'
 export { PrismaClientTsGenerator } from './generator'
 export { getDMMF } from './getDMMF'
