+++ packages/client/src/generation/generator.ts
@@ -6,10 +6,11 @@
 
 import path from 'node:path'
 
-import { dmmfToTypes, externalToInternalDmmf, PrismaClientJsGenerator } from '@prisma/client-generator-js'
+import { PrismaClientJsGenerator } from '@prisma/client-generator-js'
 import { generatorHandler } from '@prisma/generator-helper'
 
-export { dmmfToTypes, externalToInternalDmmf }
+export { dmmfToTypes } from '@prisma/client-generator-js'
+export { externalToInternalDmmf } from '@prisma/internals'
 
 if (process.argv[1] === __filename) {
   const generator = new PrismaClientJsGenerator({
