+++ packages/get-dmmf/README.md
@@ -1,0 +1,3 @@
+# @prisma/get-dmmf
+
+This package provides a function to get the DMMF (Data Model Meta Format) from a Prisma schema. The DMMF is a JSON representation of the Prisma schema, which can be used for various purposes such as generating code or creating custom tools.
