+++ packages/internals/src/engine-commands/getConfig.ts
@@ -1,5 +1,6 @@
 import Debug from '@prisma/debug'
 import type { DataSource, GeneratorConfig } from '@prisma/generator'
+import { SchemaFileInput } from '@prisma/get-dmmf'
 import { getBinaryTargetForCurrentPlatform } from '@prisma/get-platform'
 import * as E from 'fp-ts/Either'
 import { pipe } from 'fp-ts/lib/function'
@@ -7,7 +8,6 @@ import { bold, red } from 'kleur/colors'
 import { match } from 'ts-pattern'
 
 import { ErrorArea, getWasmError, isWasmPanic, RustPanic, WasmPanic } from '../panic'
-import { type SchemaFileInput } from '../utils/schemaFileInput'
 import { prismaSchemaWasm } from '../wasm'
 import { addVersionDetailsToErrorMessage } from './errorHelpers'
 import {
