+++ packages/internals/src/engine-commands/index.ts
@@ -2,7 +2,7 @@ export { formatSchema } from './formatSchema'
 export type { ConfigMetaFormat } from './getConfig'
 export { getConfig } from './getConfig'
 export type { GetDMMFOptions } from './getDmmf'
-export { getDMMF } from './getDmmf'
+export { externalToInternalDmmf, getDMMF, getInternalDMMF } from './getDmmf'
 export { getEnginesInfo, resolveEngine } from './getEnginesInfo'
 export { getEngineVersion } from './getEngineVersion'
 export { getLintWarningsAsText, handleLintPanic, lintSchema } from './lintSchema'
