+++ packages/internals/src/index.ts
@@ -97,8 +97,7 @@ export {
   PRISMA_POSTGRES_PROTOCOL,
   PRISMA_POSTGRES_PROVIDER,
 } from './utils/prismaPostgres'
-export { extractSchemaContent, type SchemaFileInput } from './utils/schemaFileInput'
-export { type MultipleSchemas } from './utils/schemaFileInput'
+export { extractSchemaContent } from './utils/schemaFileInput'
 export { setClassName } from './utils/setClassName'
 export { toSchemasContainer, toSchemasWithConfigDir } from './utils/toSchemasContainer'
 export {
@@ -109,6 +108,7 @@ export { vercelPkgPathRegex } from './utils/vercelPkgPathRegex'
 export { warnOnce } from './warnOnce'
 export * as wasm from './wasm'
 export { wasmSchemaEngineLoader } from './WasmSchemaEngineLoader'
+export { type MultipleSchemas, type SchemaFileInput } from '@prisma/get-dmmf'
 export type { BinaryTarget } from '@prisma/get-platform'
 export { getBinaryTargetForCurrentPlatform } from '@prisma/get-platform'
 export type { GetSchemaResult } from '@prisma/schema-files-loader'
