+++ packages/internals/src/utils/schemaFileInput.ts
@@ -1,7 +1,4 @@
-export type SingleSchema = string
-export type MultipleSchemaTuple = [filename: string, content: string]
-export type MultipleSchemas = Array<MultipleSchemaTuple>
-export type SchemaFileInput = SingleSchema | MultipleSchemas
+import { MultipleSchemas } from '@prisma/get-dmmf'
 
 export function extractSchemaContent(multipleSchemas: MultipleSchemas): string[] {
   return multipleSchemas.map(([, content]) => content)
