+++ packages/internals/src/utils/toSchemasContainer.ts
@@ -1,6 +1,7 @@
+import { MultipleSchemas } from '@prisma/get-dmmf'
+
 import { SchemaContext } from '../cli/schemaContext'
 import { MigrateTypes } from '../migrateTypes'
-import { MultipleSchemas } from './schemaFileInput'
 
 export function toSchemasContainer(schemas: MultipleSchemas): MigrateTypes.SchemasContainer {
   return {
